# Optimizations for history copying.